pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.75.0",
    "psycopg2-binary>=2.9.10",
    "flask-wtf>=1.2.2",
    "oauthlib>=3.2.2",
    "anthropic>=0.49.0",
    "google-generativeai>=0.8.4",
]
